fixed force delete scope

+++ src/masoniteorm/scopes/SoftDeleteScope.py
@@ -34,7 +34,9 @@ class SoftDeleteScope(BaseScope):
         builder.remove_global_scope("_where_null", action="select")
         return builder.where_not_null(self.deleted_at_column)
 
-    def _force_delete(self, model, builder):
+    def _force_delete(self, model, builder, query=False):
+        if query:
+            return builder.remove_global_scope(self).set_action("delete").to_sql()  
         return builder.remove_global_scope(self).delete()
 
     def _restore(self, model, builder):
